app/requirements.txt
django==3.2
djangorestframework
django-cors-headers
psycopg2